Added valuator_menu and valuator_collection as separate classes. Moved code for buttons and display area to Valuator_Menu class
