+++ tcg-valuator/tcg-valuator.pyproj
@@ -31,6 +31,8 @@
     <Compile Include="feetometers.py" />
     <Compile Include="main.py" />
     <Compile Include="valuator.py" />
+    <Compile Include="valuator_collection.py" />
+    <Compile Include="valuator_menu.py" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="images\" />
