+++ tcg-valuator/valuator_collection.py
@@ -1,0 +1,3 @@
+class Valuator_Collection:
+    def __init__(self):
+        self.collection = ""
